+++ src/Symfony/Component/JsonStreamer/CHANGELOG.md
@@ -5,6 +5,7 @@ CHANGELOG
 ---
 
  * Remove `nikic/php-parser` dependency
+ * Add `_current_object` to the context passed to value transformers during write operations
 
 7.3
 ---